requirements.txt
requests==2.28.2
Pillow==10.0.0
